fix: home signal did not return to GREEN
fix issue where home signal signal did not return to GREEN even when home signal switch was cleared on the web page.
Arduino UNO simply sends the home signal button state, it no longer calculates if the home signal switch was pressed or not. This is done on the signal unit.